+++ lib/generators/react_on_rails/generator_messages.rb
@@ -1,0 +1,39 @@
+module GeneratorMessages
+  class << self
+    def output
+      @output ||= []
+    end
+
+    def add_error(message)
+      output << format_error(message)
+    end
+
+    def add_warning(message)
+      output << format_warning(message)
+    end
+
+    def add_info(message)
+      output << format_info(message)
+    end
+
+    def errors
+      output
+    end
+
+    def format_error(msg)
+      Rainbow("ERROR: #{msg}").red
+    end
+
+    def format_warning(msg)
+      Rainbow("WARNING: #{msg}").orange
+    end
+
+    def format_info(msg)
+      Rainbow("#{msg}").green
+    end
+
+    def clear
+      @output = []
+    end
+  end
+end