+++ spec/dummy-react-013/client/app/reducers/reducersIndex.jsx
@@ -1,5 +1,0 @@
-import helloWorldReducer from './HelloWorldReducer';
-
-// This is how you do a directory of reducers.
-// The `import * as reducers` does not work for a directory, but only with a single file
-export default { helloWorldData: helloWorldReducer };